Create automated_zone_transfer.sh
This is a simple script designed to automate DNS zone transfer brute forcing for a given domain. Rather than having to run multiple host commands with different arguments, you just need to type in a domain and the script automatically tries to perform a zone transfer on all available nameservers.